feat(ui): serve the app and the account chooser from one root route
With the chooser on the root route, /home was a second landing page and the
two redirected into each other. The root now renders the app (apps / drives
/ account tabs) whenever a session account exists and the chooser otherwise,
so opening the app while signed in goes straight in, picking an account
flips in place, and the post-setup Continue buttons just point at /.
